# contracts/openapi/catalog.json
{
    "openapi": "3.0.4",
    "info": {
        "title": "LocalizeStay Catalog API",
        "version": "v1"
    },
    "paths": {
        "/v1/properties": {
            "post": {
                "tags": [
                    "Properties"
                ],
                "operationId": "createProperty",
                "parameters": [
                    {
                        "name": "X-Host-Reference-Id",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreatePropertyRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "headers": {
                            "Location": {
                                "description": "Caminho versionado da Property cadastrada.",
                                "schema": {
                                    "type": "string"
                                }
                            }
                        },
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Property"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/problem+json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetails"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/problem+json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetails"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/v1/properties/{propertyId}": {
            "patch": {
                "tags": [
                    "Properties"
                ],
                "operationId": "updateProperty",
                "parameters": [
                    {
                        "name": "propertyId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    },
                    {
                        "name": "X-Host-Reference-Id",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdatePropertyRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Property"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/problem+json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetails"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/problem+json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetails"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/problem+json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetails"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/problem+json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetails"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "CreatePropertyRequest": {
                "required": [
                    "location",
                    "name"
                ],
                "type": "object",
                "properties": {
                    "name": {
                        "maxLength": 120,
                        "minLength": 1,
                        "pattern": ".*\\S.*",
                        "type": "string"
                    },
                    "location": {
                        "maxLength": 500,
                        "minLength": 1,
                        "pattern": ".*\\S.*",
                        "type": "string"
                    }
                },
                "additionalProperties": false
            },
            "ProblemDetailItem": {
                "required": [
                    "field",
                    "message"
                ],
                "type": "object",
                "properties": {
                    "field": {
                        "type": "string",
                        "nullable": true
                    },
                    "message": {
                        "type": "string",
                        "nullable": true
                    }
                },
                "additionalProperties": false
            },
            "ProblemDetails": {
                "required": [
                    "code",
                    "detail",
                    "details",
                    "instance",
                    "status",
                    "title",
                    "traceId",
                    "type"
                ],
                "type": "object",
                "properties": {
                    "type": {
                        "type": "string",
                        "nullable": true
                    },
                    "title": {
                        "type": "string",
                        "nullable": true
                    },
                    "status": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "detail": {
                        "type": "string",
                        "nullable": true
                    },
                    "instance": {
                        "type": "string",
                        "nullable": true
                    },
                    "code": {
                        "type": "string",
                        "nullable": true
                    },
                    "details": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ProblemDetailItem"
                        },
                        "nullable": true
                    },
                    "traceId": {
                        "type": "string",
                        "nullable": true
                    }
                },
                "additionalProperties": false
            },
            "Property": {
                "required": [
                    "hostReferenceId",
                    "id",
                    "location",
                    "name",
                    "status"
                ],
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "name": {
                        "maxLength": 120,
                        "minLength": 1,
                        "pattern": ".*\\S.*",
                        "type": "string"
                    },
                    "location": {
                        "maxLength": 500,
                        "minLength": 1,
                        "pattern": ".*\\S.*",
                        "type": "string"
                    },
                    "hostReferenceId": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "status": {
                        "type": "string"
                    }
                },
                "additionalProperties": false
            },
            "UpdatePropertyRequest": {
                "minProperties": 1,
                "type": "object",
                "properties": {
                    "name": {
                        "maxLength": 120,
                        "minLength": 1,
                        "pattern": ".*\\S.*",
                        "type": "string"
                    },
                    "location": {
                        "maxLength": 500,
                        "minLength": 1,
                        "pattern": ".*\\S.*",
                        "type": "string"
                    }
                },
                "additionalProperties": false
            }
        }
    },
    "tags": [
        {
            "name": "Properties"
        }
    ]
}